Updated e2fsprogs to 1.42.7.
This new version resolves many issues, and is also more memory-efficient on large filesystems.

+++ release/src/router/e2fsprogs/lib/ext2fs/link.c
@@ -9,6 +9,7 @@
  * %End-Header%
  */
 
+#include "config.h"
 #include <stdio.h>
 #include <string.h>
 #if HAVE_UNISTD_H
@@ -52,9 +53,9 @@ static int link_proc(struct ext2_dir_entry *dirent,
 	 * if so, absorb it into this one.
 	 */
 	next = (struct ext2_dir_entry *) (buf + offset + curr_rec_len);
-	if ((offset + curr_rec_len < blocksize - 8) &&
+	if ((offset + (int) curr_rec_len < blocksize - 8) &&
 	    (next->inode == 0) &&
-	    (offset + curr_rec_len + next->rec_len <= blocksize)) {
+	    (offset + (int) curr_rec_len + (int) next->rec_len <= blocksize)) {
 		curr_rec_len += next->rec_len;
 		ls->err = ext2fs_set_rec_len(ls->fs, curr_rec_len, dirent);
 		if (ls->err)
